back/requirements.txt
asgiref==3.2.3
autopep8==1.4.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.1
django-cors-headers==3.2.0
django-import-export==2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pytz==2019.3
PyYAML==5.2
sqlparse==0.3.0
tablib==0.14.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
